+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased 1.21.11]
+## [1.21.11-4.1.0] - 2026-01-16
+Port to 1.21.11! Not thoroughly tested yet. Expect issues.
 ### Changed
 - Spears now work like vanilla spears
 - Spear Tip renamed to Spearhead
+- Netherite coating now adds 40 to rarity
+### Known Issues
+- Some item models are not working as intended.
 
 ## [1.21.1-4.0.30] - 2025-11-23
 ### Added